requirements.txt
pywin32>=220 ; sys_platform == 'win32'
lxml>=3.3.5
Scrapy>=1.1.0
PyMySQL>=0.7.9
psycopg2-binary>=2.8.4
hjson>=1.5.8
elasticsearch>=2.4
beautifulsoup4>=4.3.2
readability-lxml>=0.6.2
newspaper3k>=0.2.8
langdetect>=1.0.7
python-dateutil>=2.4.0
plac>=0.9.6
PyDispatcher>=2.0.5
dotmap>=1.2.17
warcio>=1.3.3
ago>=0.0.9
six>=1.10.0
hurry.filesize>=0.9
bs4
faust-cchardet>=2.1.18
boto3
markdownify
